requirements.txt
aws-cdk-lib
aws_cdk.aws_lambda_python_alpha
constructs>=10.0.0,<11.0.0
jupyter
pre-commit
yamldataclassconfig
boto3
